Revert to previous deletion method
And addressed raveens comments
